[LG-80] Refactor commons dto
