+++ commons/src/main/java/dto/business/product/ManufacturerResponseDto.java
@@ -1,4 +1,4 @@
-package dto.business;
+package dto.business.product;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
